pyproject.toml
[project]
name = "zommalabskg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.63.0",
    "docling-core>=2.52.0",
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
    "langchain-voyageai>=0.3.0",
    "langchain-google-genai>=1.0.0",
    "langchain-xai>=0.2.0",
    "langgraph>=0.0.10",
    "neo4j>=6.0.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.1",
    "rapidfuzz>=3.14.3",
    "rdflib>=7.5.0",
    "typing-extensions>=4.15.0",
    "fastmcp>=2.14.1",
    "langchain-mcp-adapters>=0.1.0",
    "langgraph-cli[inmem]>=0.2.0",
    "langchain-community>=0.4.1",
    "langchain-anthropic>=1.3.0",
    "google-generativeai>=0.8.6",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
